test(canvas): share Skia pixel sampler
Fixes #2462.
